move some methods from Keen::Client to Keen:Helpers move config variable to a config class
