+++ lib/keen.rb
@@ -26,12 +26,6 @@ module Keen
     extend Forwardable
 
     def_delegators :default_client,
-                   :project_id, :project_id=,
-                   :write_key, :write_key=,
-                   :read_key, :read_key=,
-                   :api_url, :api_url=
-
-    def_delegators :default_client,
                    :publish, :publish_async, :publish_batch,
                    :beacon_url
 
@@ -51,6 +45,10 @@ module Keen
         logger.level = Logger::INFO
         logger
       }.call
+    end
+
+    def config
+      default_client.config
     end
 
     private
