+++ lib/keen/client/maintenance_methods.rb
@@ -16,9 +16,9 @@ module Keen
         query_params = preprocess_params(params) if params != {}
 
         begin
-          response = Keen::HTTP::Sync.new(self.api_url).delete(
+          response = Keen::HTTP::Sync.new(config.api_url).delete(
               :path => [api_event_collection_resource_path(event_collection), query_params].compact.join('?'),
-              :headers => api_headers(self.master_key, "sync"))
+              :headers => config.api_headers(config.master_key, "sync"))
         rescue Exception => http_error
           raise HttpError.new("Couldn't perform delete of #{event_collection} on Keen IO: #{http_error.message}", http_error)
         end
